Chaning utlGeoKm and utlLonLat to double precision
 * closes #143, which was due to underflow (thanks to Martin van Driel)
 * extensions need to be recompiled

+++ obspy.signal/trunk/obspy/signal/src/coordtrans.c
@@ -9,9 +9,9 @@
 #define DRLT	9.9330647e-1
 
 
-void utl_geo_km(float orig_lon, float orig_lat, float rota, float *lon, float *lat) {
-   float	olon;
-   float	olat;
+void utl_geo_km(double orig_lon, double orig_lat, double rota, double *lon, double *lat) {
+   double	olon;
+   double	olat;
    double	lat_fac;	/* conversion factor for latitude in km */
    double	lon_fac;	/* conversion factor for longitude in km */
    double	snr;		/* sin of rotation angle */
@@ -20,8 +20,8 @@ void utl_geo_km(float orig_lon, float orig_lat, float rota, float *lon, float *l
    double	dlt2;
    double	del;
    double	radius;
-   float	tmp;
-   float	tmp_x, tmp_y;
+   double	tmp;
+   double	tmp_x, tmp_y;
 
    /* convert everything to minutes */
    orig_lat = 60.0f * orig_lat;
@@ -30,8 +30,8 @@ void utl_geo_km(float orig_lon, float orig_lat, float rota, float *lon, float *l
    olat = orig_lat;
 
    /* latitude */
-   dlt1 = atan(DRLT * tan((double)olat * DRAD/60.0));
-   dlt2 = atan(DRLT * tan(((double)olat +1.0) * DRAD/60.0));
+   dlt1 = atan(DRLT * tan(olat * DRAD/60.0));
+   dlt2 = atan(DRLT * tan((olat +1.0) * DRAD/60.0));
    del  = dlt2 - dlt1;
    radius = E_RAD * (1.0 - (sin(dlt1)*sin(dlt1) / E_FLAT));
    lat_fac = radius * del;
@@ -42,8 +42,8 @@ void utl_geo_km(float orig_lon, float orig_lat, float rota, float *lon, float *l
    lon_fac = dlt2 / cos(dlt1);
 
    /* rotation */
-   snr = sin((double)rota * DRAD);
-   csr = cos((double)rota * DRAD);
+   snr = sin(rota * DRAD);
+   csr = cos(rota * DRAD);
 
 
    *lat *= 60.0f; 
@@ -52,17 +52,17 @@ void utl_geo_km(float orig_lon, float orig_lat, float rota, float *lon, float *l
    tmp_x = (*lon) - orig_lon;
    tmp_y = (*lat) - orig_lat;
 
-   tmp   = (float)(atan(DRLT * tan(DRAD * ((*lat)+orig_lat)/120.0)));
-   tmp_x = (float)((double)tmp_x * lon_fac * cos(tmp));
-   tmp_y = (float)((double)tmp_y * lat_fac);
+   tmp   = atan(DRLT * tan(DRAD * ((*lat)+orig_lat)/120.0));
+   tmp_x = tmp_x * lon_fac * cos(tmp);
+   tmp_y = tmp_y * lat_fac;
 
-   *lon = (float)(csr*tmp_x - snr*tmp_y);
-   *lat = (float)(csr*tmp_y + snr*tmp_x);
+   *lon = csr*tmp_x - snr*tmp_y;
+   *lat = csr*tmp_y + snr*tmp_x;
 }
 
-void utl_lonlat(float orig_lon,float orig_lat,float x,float y,float *lon,float *lat) {
-   float        olon;
-   float        olat;
+void utl_lonlat(double orig_lon,double orig_lat,double x,double y,double *lon,double *lat) {
+   double        olon;
+   double        olat;
    double       lat_fac;        /* conversion factor for latitude in km */
    double       lon_fac;        /* conversion factor for longitude in km */
    double       snr;            /* sin of rotation angle */
@@ -71,9 +71,9 @@ void utl_lonlat(float orig_lon,float orig_lat,float x,float y,float *lon,float *
    double       dlt2;
    double       del;
    double       radius;
-   float        tmp;
-   float        tmp_x, tmp_y;
-   float 	rota=0.0;
+   double        tmp;
+   double        tmp_x, tmp_y;
+   double 	rota=0.0;
 
    /* convert everything to minutes */
    orig_lat = 60.0f * orig_lat;
@@ -82,8 +82,8 @@ void utl_lonlat(float orig_lon,float orig_lat,float x,float y,float *lon,float *
    olat = orig_lat;
 
    /* latitude */
-   dlt1 = atan(DRLT * tan((double)olat * DRAD/60.0));
-   dlt2 = atan(DRLT * tan(((double)olat +1.0) * DRAD/60.0));
+   dlt1 = atan(DRLT * tan(olat * DRAD/60.0));
+   dlt2 = atan(DRLT * tan((olat +1.0) * DRAD/60.0));
    del  = dlt2 - dlt1;
    radius = E_RAD * (1.0 - (sin(dlt1)*sin(dlt1) / E_FLAT));
    lat_fac = radius * del;
@@ -94,18 +94,18 @@ void utl_lonlat(float orig_lon,float orig_lat,float x,float y,float *lon,float *
    lon_fac = dlt2 / cos(dlt1);
 
    /* rotation */
-   snr = sin((double)rota * DRAD);
-   csr = cos((double)rota * DRAD);
+   snr = sin(rota * DRAD);
+   csr = cos(rota * DRAD);
 
 
-    tmp_x = (float)(snr*y + csr*x);
-    tmp_y = (float)(csr*y - snr*x);
+    tmp_x = snr*y + csr*x;
+    tmp_y = csr*y - snr*x;
 
-    tmp_y = (float)(tmp_y/lat_fac);
+    tmp_y = tmp_y/lat_fac;
     tmp_y += olat;
 
-    tmp = (float)(atan(DRLT * tan(DRAD * (tmp_y+orig_lat)/120.0)));
-    tmp_x = (float)(tmp_x / (lon_fac * cos(tmp)));
+    tmp = atan(DRLT * tan(DRAD * (tmp_y+orig_lat)/120.0));
+    tmp_x = tmp_x / (lon_fac * cos(tmp));
     tmp_x += olon;
 
     *lon = tmp_x/60.0f;
